+++ config/local.sh
@@ -6,12 +6,15 @@
 # Be sure to "source config.sh" each time 
 #-----------------------------------------------------------------
 
+# Projects path
 export QTRADE=$HOME/dev/projects/ppQuanTrade
 export ZIPLINE=$HOME/dev/projects/zipline
-export QTRADEDATA=$QTRADE/database
+export QTRADEDATA=$QTRADE/data
 export QTRADE_SQLITE=stocks.db
 export QTRADE_LOG=$QTRADE/quantrade.log
 export QTRADE_CONFIGURATION=local
+
+# For development: add zipline and neuronquant lib in python path
 export PYTHONPATH=$PYTHONPATH:$QTRADE:$ZIPLINE
 
 #NotifyMyAndroid key, check http://www.notifymyandroid.com/index.jsp