+++ neuronquant/utils/db_utils.py
@@ -1,0 +1,52 @@
+#
+# Copyright 2012 Xavier Bruhiere
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+#TODO: Same for google json and xml retrievig
+googleCode = dict()
+
+#TODO: codes for non-us markets are different. Like JXR.PA for archos
+#TODO: Tester, comprendre, et reformater ces valeurs
+yahooCode = {'ask': 'a'                          , 'average daily volume': 'a2'            , 'ask size': 'a5'                  ,
+        'bid': 'b'                               , 'ask rt': 'b2'                          , 'bid rt': 'b3'                    , 'dividend yield': 'y' ,
+        'book value': 'b4'                       , 'bid size': 'b6'                        , 'change and percent': 'c'         ,
+        'change': 'c1'                           , 'commission': 'c3'                      , 'change rt': 'c6'                 ,
+        'after hours change rt': 'c8'            , 'dividend': 'd'                         , 'last trade date': 'd1'           ,
+        'trade date': 'd2'                       , 'earnings': 'e'                         , 'error': 'e1'                     ,
+        'eps estimate year': 'e7'                , 'eps estimate next year': 'e8'          , 'eps estimate next quarter': 'e9' ,
+        'float shares': 'f6'                     , 'day low': 'g'                          , 'day high': 'h'                   ,
+        '52-week low': 'j'                       , '52-week high': 'k'                     , 'holdings gain percent': 'g1'     ,
+        'annualized gain': 'g3'                  , 'holdings gain': 'g4'                   , 'holdings gain percent rt': 'g5'  ,
+        'holdings gain rt': 'g6'                 , 'more info': 'i'                        , 'order book rt': 'i5'             ,
+        'market capitalization': 'j1'            , 'market cap rt': 'j3'                   , 'EBITDA': 'j4'                    ,
+        'change from 52-week': 'j5'              , 'percent change from 52-week low': 'j6' , 'last trade rt with time': 'k1'   ,
+        'change percent rt': 'k2'                , 'last trade size': 'k3'                 , 'change from 52-week high': 'k4'  ,
+        'percebt change from 52-week high': 'k5' , 'last trade with time': 'l'             , 'last trade price': 'l1'          ,
+        'high limit': 'l2'                       , 'low limit': 'l3'                       , 'day range': 'm'                  ,
+        'day range rt': 'm2'                     , '50-day ma': 'm3'                       , '200-day ma': 'm4'                ,
+        'percent change from 50-day ma': 'm8'    , 'name': 'n'                             , 'notes': 'n4'                     ,
+        'open': 'o'                              , 'previous close': 'p'                   , 'price paid': 'p1'                ,
+        'change percent': 'p2'                   , 'price/sales': 'p5'                     , 'price/book': 'p6'                ,
+        'ex-dividend date': 'q'                  , 'p/e ratio': 'r'                        , 'dividend date': 'r1'             ,
+        'p/e ratio rt': 'r2'                     , 'peg ratio': 'r5'                       , 'price/eps estimate year': 'r6'   ,
+        'price/eps estimate next year': 'r7'     , 'symbol': 's'                           , 'shares owned': 's1'              ,
+        'short ratio': 's7'                      , 'last trade time': 't1'                 , 'trade links': 't6'               ,
+        'ticker trend': 't7'                     , '1 year target price': 't8'             , 'volume': 'v'                     ,
+        'holdings value': 'v1'                   , 'holdings value rt': 'v7'               , '52-week range': 'w'              ,
+        'day value change': 'w1'                 , 'day value change rt': 'w4'             , 'stock exchange': 'x'}
+
+
+class Fields:
+    QUOTES = ['open', 'low', 'high', 'close', 'volume', 'adj_close']